Reorganise test package
The main package was reorganised in https://github.com/AY2122S1-CS2103T-W16-2/tp/pull/34.
Reorganised the test package to mirror changes made in the earlier PR.
